+++ apps/data/road_segmenting/example_roadnet_to_db.py
@@ -3,6 +3,7 @@ import os
 import requests
 
 from road_fetcher import vegnet_to_geojson
+from road_filter import filter_road
 
 municipality = 5001
 type_of_road = "kg"
@@ -12,37 +13,6 @@ min_segments = 2
 # Credentials for connecting and writing to the API
 API_username = os.environ["API_USERNAME"]
 API_password = os.environ["API_PASSWORD"]
-
-
-def filter_road(road):
-    """
-    Filteres the input road segment to satisfy DB layout.
-    :param road: A dictionary containing a road segment
-    :return: A filtered dict with the road segment
-    """
-    # Format the linestring correctly
-    linestring = ""
-    for pair in road["geometry"]["coordinates"]:
-        linestring += str(pair[0]) + " " + str(pair[1]) + ","
-    linestring = linestring.rstrip(",")
-    geometry = "SRID={};LINESTRING({})".format(road["properties"]["geometri"]["srid"], linestring)
-
-    filtered_road = {
-        "the_geom":        geometry,
-        "county":          road["properties"]["fylke"],
-        "href":            road["properties"]["href"],
-        "category":        road["properties"]["kategori"],
-        "municipality":    road["properties"]["kommune"],
-        "startdate":       road["properties"]["metadata"]["startdato"],
-        "region":          road["properties"]["region"],
-        "status":          road["properties"]["status"],
-        "stretchdistance": road["properties"]["strekningslengde"],
-        "typeofroad":      road["properties"]["typeVeg"],
-        "roadsectionid":   road["properties"]["veglenkeid"],
-        "vrefshortform":   road["properties"]["vrefkortform"],
-    }
-
-    return filtered_road
 
 
 def format_to_db(municipality, type_road):
